separate channel list components
