+++ src/components/Main/Navigation/NavigationContent/Channels.vue
@@ -6,7 +6,7 @@
 import { defineComponent, reactive, computed } from '@vue/composition-api'
 import store from '@/store'
 import { makeStyles } from '@/lib/styles'
-import ChannelList from '@/components/Main/Navigation/ChannelList.vue'
+import ChannelList from '@/components/Main/Navigation/ChannelList/ChannelList.vue'
 
 type Props = {}
