+++ src/components/Main/Navigation/NavigationContent/Home.vue
@@ -17,7 +17,7 @@ import { defineComponent, reactive, computed } from '@vue/composition-api'
 import store from '@/store'
 import { makeStyles } from '@/lib/styles'
 import EmptyState from '@/components/UI/EmptyState.vue'
-import ChannelList from '@/components/Main/Navigation/ChannelList.vue'
+import ChannelList from '@/components/Main/Navigation/ChannelList/ChannelList.vue'
 
 type Props = {}
